--- dcrjson/btcdextcmds_test.go
@@ -1,206 +1,0 @@
-// Copyright (c) 2014 The btcsuite developers
-// Copyright (c) 2015-2016 The Decred developers
-// Use of this source code is governed by an ISC
-// license that can be found in the LICENSE file.
-
-package dcrjson_test
-
-import (
-	"bytes"
-	"encoding/json"
-	"fmt"
-	"reflect"
-	"testing"
-
-	"github.com/decred/dcrd/dcrjson"
-)
-
-// TestBtcdExtCmds tests all of the btcd extended commands marshal and unmarshal
-// into valid results include handling of optional fields being omitted in the
-// marshalled command, while optional fields with defaults have the default
-// assigned on unmarshalled commands.
-func TestBtcdExtCmds(t *testing.T) {
-	t.Parallel()
-
-	testID := int(1)
-	tests := []struct {
-		name         string
-		newCmd       func() (interface{}, error)
-		staticCmd    func() interface{}
-		marshalled   string
-		unmarshalled interface{}
-	}{
-		{
-			name: "debuglevel",
-			newCmd: func() (interface{}, error) {
-				return dcrjson.NewCmd("debuglevel", "trace")
-			},
-			staticCmd: func() interface{} {
-				return dcrjson.NewDebugLevelCmd("trace")
-			},
-			marshalled: `{"jsonrpc":"1.0","method":"debuglevel","params":["trace"],"id":1}`,
-			unmarshalled: &dcrjson.DebugLevelCmd{
-				LevelSpec: "trace",
-			},
-		},
-		{
-			name: "node",
-			newCmd: func() (interface{}, error) {
-				return dcrjson.NewCmd("node", dcrjson.NRemove, "1.1.1.1")
-			},
-			staticCmd: func() interface{} {
-				return dcrjson.NewNodeCmd("remove", "1.1.1.1", nil)
-			},
-			marshalled: `{"jsonrpc":"1.0","method":"node","params":["remove","1.1.1.1"],"id":1}`,
-			unmarshalled: &dcrjson.NodeCmd{
-				SubCmd: dcrjson.NRemove,
-				Target: "1.1.1.1",
-			},
-		},
-		{
-			name: "node",
-			newCmd: func() (interface{}, error) {
-				return dcrjson.NewCmd("node", dcrjson.NDisconnect, "1.1.1.1")
-			},
-			staticCmd: func() interface{} {
-				return dcrjson.NewNodeCmd("disconnect", "1.1.1.1", nil)
-			},
-			marshalled: `{"jsonrpc":"1.0","method":"node","params":["disconnect","1.1.1.1"],"id":1}`,
-			unmarshalled: &dcrjson.NodeCmd{
-				SubCmd: dcrjson.NDisconnect,
-				Target: "1.1.1.1",
-			},
-		},
-		{
-			name: "node",
-			newCmd: func() (interface{}, error) {
-				return dcrjson.NewCmd("node", dcrjson.NConnect, "1.1.1.1", "perm")
-			},
-			staticCmd: func() interface{} {
-				return dcrjson.NewNodeCmd("connect", "1.1.1.1", dcrjson.String("perm"))
-			},
-			marshalled: `{"jsonrpc":"1.0","method":"node","params":["connect","1.1.1.1","perm"],"id":1}`,
-			unmarshalled: &dcrjson.NodeCmd{
-				SubCmd:        dcrjson.NConnect,
-				Target:        "1.1.1.1",
-				ConnectSubCmd: dcrjson.String("perm"),
-			},
-		},
-		{
-			name: "node",
-			newCmd: func() (interface{}, error) {
-				return dcrjson.NewCmd("node", dcrjson.NConnect, "1.1.1.1", "temp")
-			},
-			staticCmd: func() interface{} {
-				return dcrjson.NewNodeCmd("connect", "1.1.1.1", dcrjson.String("temp"))
-			},
-			marshalled: `{"jsonrpc":"1.0","method":"node","params":["connect","1.1.1.1","temp"],"id":1}`,
-			unmarshalled: &dcrjson.NodeCmd{
-				SubCmd:        dcrjson.NConnect,
-				Target:        "1.1.1.1",
-				ConnectSubCmd: dcrjson.String("temp"),
-			},
-		},
-		{
-			name: "generate",
-			newCmd: func() (interface{}, error) {
-				return dcrjson.NewCmd("generate", 1)
-			},
-			staticCmd: func() interface{} {
-				return dcrjson.NewGenerateCmd(1)
-			},
-			marshalled: `{"jsonrpc":"1.0","method":"generate","params":[1],"id":1}`,
-			unmarshalled: &dcrjson.GenerateCmd{
-				NumBlocks: 1,
-			},
-		},
-		{
-			name: "getbestblock",
-			newCmd: func() (interface{}, error) {
-				return dcrjson.NewCmd("getbestblock")
-			},
-			staticCmd: func() interface{} {
-				return dcrjson.NewGetBestBlockCmd()
-			},
-			marshalled:   `{"jsonrpc":"1.0","method":"getbestblock","params":[],"id":1}`,
-			unmarshalled: &dcrjson.GetBestBlockCmd{},
-		},
-		{
-			name: "getcurrentnet",
-			newCmd: func() (interface{}, error) {
-				return dcrjson.NewCmd("getcurrentnet")
-			},
-			staticCmd: func() interface{} {
-				return dcrjson.NewGetCurrentNetCmd()
-			},
-			marshalled:   `{"jsonrpc":"1.0","method":"getcurrentnet","params":[],"id":1}`,
-			unmarshalled: &dcrjson.GetCurrentNetCmd{},
-		},
-	}
-
-	t.Logf("Running %d tests", len(tests))
-	for i, test := range tests {
-		// Marshal the command as created by the new static command
-		// creation function.
-		marshalled, err := dcrjson.MarshalCmd("1.0", testID, test.staticCmd())
-		if err != nil {
-			t.Errorf("MarshalCmd #%d (%s) unexpected error: %v", i,
-				test.name, err)
-			continue
-		}
-
-		if !bytes.Equal(marshalled, []byte(test.marshalled)) {
-			t.Errorf("Test #%d (%s) unexpected marshalled data - "+
-				"got %s, want %s", i, test.name, marshalled,
-				test.marshalled)
-			continue
-		}
-
-		// Ensure the command is created without error via the generic
-		// new command creation function.
-		cmd, err := test.newCmd()
-		if err != nil {
-			t.Errorf("Test #%d (%s) unexpected NewCmd error: %v ",
-				i, test.name, err)
-		}
-
-		// Marshal the command as created by the generic new command
-		// creation function.
-		marshalled, err = dcrjson.MarshalCmd("1.0", testID, cmd)
-		if err != nil {
-			t.Errorf("MarshalCmd #%d (%s) unexpected error: %v", i,
-				test.name, err)
-			continue
-		}
-
-		if !bytes.Equal(marshalled, []byte(test.marshalled)) {
-			t.Errorf("Test #%d (%s) unexpected marshalled data - "+
-				"got %s, want %s", i, test.name, marshalled,
-				test.marshalled)
-			continue
-		}
-
-		var request dcrjson.Request
-		if err := json.Unmarshal(marshalled, &request); err != nil {
-			t.Errorf("Test #%d (%s) unexpected error while "+
-				"unmarshalling JSON-RPC request: %v", i,
-				test.name, err)
-			continue
-		}
-
-		cmd, err = dcrjson.UnmarshalCmd(&request)
-		if err != nil {
-			t.Errorf("UnmarshalCmd #%d (%s) unexpected error: %v", i,
-				test.name, err)
-			continue
-		}
-
-		if !reflect.DeepEqual(cmd, test.unmarshalled) {
-			t.Errorf("Test #%d (%s) unexpected unmarshalled command "+
-				"- got %s, want %s", i, test.name,
-				fmt.Sprintf("(%T) %+[1]v", cmd),
-				fmt.Sprintf("(%T) %+[1]v\n", test.unmarshalled))
-			continue
-		}
-	}
-}
