+++ dcrjson/chainsvrcmds.go
@@ -30,6 +30,22 @@ const (
 	ANOneTry AddNodeSubCmd = "onetry"
 )
 
+// NodeSubCmd defines the type used in the addnode JSON-RPC command for the
+// sub command field.
+type NodeSubCmd string
+
+const (
+	// NConnect indicates the specified host that should be connected to.
+	NConnect NodeSubCmd = "connect"
+
+	// NRemove indicates the specified peer that should be removed as a
+	// persistent peer.
+	NRemove NodeSubCmd = "remove"
+
+	// NDisconnect indicates the specified peer should be disonnected.
+	NDisconnect NodeSubCmd = "disconnect"
+)
+
 // AddNodeCmd defines the addnode JSON-RPC command.
 type AddNodeCmd struct {
 	Addr   string
@@ -73,6 +89,21 @@ func NewCreateRawTransactionCmd(inputs []TransactionInput, amounts map[string]fl
 		Amounts:  amounts,
 		LockTime: lockTime,
 		Expiry:   expiry,
+	}
+}
+
+// DebugLevelCmd defines the debuglevel JSON-RPC command.  This command is not a
+// standard Bitcoin command.  It is an extension for btcd.
+type DebugLevelCmd struct {
+	LevelSpec string
+}
+
+// NewDebugLevelCmd returns a new DebugLevelCmd which can be used to issue a
+// debuglevel JSON-RPC command.  This command is not a standard Bitcoin command.
+// It is an extension for btcd.
+func NewDebugLevelCmd(levelSpec string) *DebugLevelCmd {
+	return &DebugLevelCmd{
+		LevelSpec: levelSpec,
 	}
 }
 
@@ -144,6 +175,122 @@ func NewEstimateSmartFeeCmd(confirmations int64, mode EstimateSmartFeeMode) *Est
 	}
 }
 
+// EstimateStakeDiffCmd defines the eststakedifficulty JSON-RPC command.
+type EstimateStakeDiffCmd struct {
+	Tickets *uint32
+}
+
+// NewEstimateStakeDiffCmd defines the eststakedifficulty JSON-RPC command.
+func NewEstimateStakeDiffCmd(tickets *uint32) *EstimateStakeDiffCmd {
+	return &EstimateStakeDiffCmd{
+		Tickets: tickets,
+	}
+}
+
+// ExistsAddressCmd defines the existsaddress JSON-RPC command.
+type ExistsAddressCmd struct {
+	Address string
+}
+
+// NewExistsAddressCmd returns a new instance which can be used to issue a
+// existsaddress JSON-RPC command.
+func NewExistsAddressCmd(address string) *ExistsAddressCmd {
+	return &ExistsAddressCmd{
+		Address: address,
+	}
+}
+
+// ExistsAddressesCmd defines the existsaddresses JSON-RPC command.
+type ExistsAddressesCmd struct {
+	Addresses []string
+}
+
+// NewExistsAddressesCmd returns a new instance which can be used to issue an
+// existsaddresses JSON-RPC command.
+func NewExistsAddressesCmd(addresses []string) *ExistsAddressesCmd {
+	return &ExistsAddressesCmd{
+		Addresses: addresses,
+	}
+}
+
+// ExistsMissedTicketsCmd defines the existsmissedtickets JSON-RPC command.
+type ExistsMissedTicketsCmd struct {
+	TxHashBlob string
+}
+
+// NewExistsMissedTicketsCmd returns a new instance which can be used to issue an
+// existsmissedtickets JSON-RPC command.
+func NewExistsMissedTicketsCmd(txHashBlob string) *ExistsMissedTicketsCmd {
+	return &ExistsMissedTicketsCmd{
+		TxHashBlob: txHashBlob,
+	}
+}
+
+// ExistsExpiredTicketsCmd defines the existsexpiredtickets JSON-RPC command.
+type ExistsExpiredTicketsCmd struct {
+	TxHashBlob string
+}
+
+// NewExistsExpiredTicketsCmd returns a new instance which can be used to issue an
+// existsexpiredtickets JSON-RPC command.
+func NewExistsExpiredTicketsCmd(txHashBlob string) *ExistsExpiredTicketsCmd {
+	return &ExistsExpiredTicketsCmd{
+		TxHashBlob: txHashBlob,
+	}
+}
+
+// ExistsLiveTicketCmd defines the existsliveticket JSON-RPC command.
+type ExistsLiveTicketCmd struct {
+	TxHash string
+}
+
+// NewExistsLiveTicketCmd returns a new instance which can be used to issue an
+// existsliveticket JSON-RPC command.
+func NewExistsLiveTicketCmd(txHash string) *ExistsLiveTicketCmd {
+	return &ExistsLiveTicketCmd{
+		TxHash: txHash,
+	}
+}
+
+// ExistsLiveTicketsCmd defines the existslivetickets JSON-RPC command.
+type ExistsLiveTicketsCmd struct {
+	TxHashBlob string
+}
+
+// NewExistsLiveTicketsCmd returns a new instance which can be used to issue an
+// existslivetickets JSON-RPC command.
+func NewExistsLiveTicketsCmd(txHashBlob string) *ExistsLiveTicketsCmd {
+	return &ExistsLiveTicketsCmd{
+		TxHashBlob: txHashBlob,
+	}
+}
+
+// ExistsMempoolTxsCmd defines the existsmempooltxs JSON-RPC command.
+type ExistsMempoolTxsCmd struct {
+	TxHashBlob string
+}
+
+// NewExistsMempoolTxsCmd returns a new instance which can be used to issue an
+// existslivetickets JSON-RPC command.
+func NewExistsMempoolTxsCmd(txHashBlob string) *ExistsMempoolTxsCmd {
+	return &ExistsMempoolTxsCmd{
+		TxHashBlob: txHashBlob,
+	}
+}
+
+// GenerateCmd defines the generate JSON-RPC command.
+type GenerateCmd struct {
+	NumBlocks uint32
+}
+
+// NewGenerateCmd returns a new instance which can be used to issue a generate
+// JSON-RPC command.
+func NewGenerateCmd(numBlocks uint32) *GenerateCmd {
+	return &GenerateCmd{
+		NumBlocks: numBlocks,
+	}
+}
+
 // GetAddedNodeInfoCmd defines the getaddednodeinfo JSON-RPC command.
 type GetAddedNodeInfoCmd struct {
 	DNS  bool
@@ -160,6 +307,15 @@ func NewGetAddedNodeInfoCmd(dns bool, node *string) *GetAddedNodeInfoCmd {
 		DNS:  dns,
 		Node: node,
 	}
+}
+
+// GetBestBlockCmd defines the getbestblock JSON-RPC command.
+type GetBestBlockCmd struct{}
+
+// NewGetBestBlockCmd returns a new instance which can be used to issue a
+// getbestblock JSON-RPC command.
+func NewGetBestBlockCmd() *GetBestBlockCmd {
+	return &GetBestBlockCmd{}
 }
 
 // GetBestBlockHashCmd defines the getbestblockhash JSON-RPC command.
@@ -379,6 +535,15 @@ func NewGetChainTipsCmd() *GetChainTipsCmd {
 	return &GetChainTipsCmd{}
 }
 
+// GetCoinSupplyCmd defines the getcoinsupply JSON-RPC command.
+type GetCoinSupplyCmd struct{}
+
+// NewGetCoinSupplyCmd returns a new instance which can be used to issue a
+// getcoinsupply JSON-RPC command.
+func NewGetCoinSupplyCmd() *GetCoinSupplyCmd {
+	return &GetCoinSupplyCmd{}
+}
+
 // GetConnectionCountCmd defines the getconnectioncount JSON-RPC command.
 type GetConnectionCountCmd struct{}
 
@@ -386,6 +551,15 @@ type GetConnectionCountCmd struct{}
 // getconnectioncount JSON-RPC command.
 func NewGetConnectionCountCmd() *GetConnectionCountCmd {
 	return &GetConnectionCountCmd{}
+}
+
+// GetCurrentNetCmd defines the getcurrentnet JSON-RPC command.
+type GetCurrentNetCmd struct{}
+
+// NewGetCurrentNetCmd returns a new instance which can be used to issue a
+// getcurrentnet JSON-RPC command.
+func NewGetCurrentNetCmd() *GetCurrentNetCmd {
+	return &GetCurrentNetCmd{}
 }
 
 // GetDifficultyCmd defines the getdifficulty JSON-RPC command.
@@ -562,6 +736,55 @@ func NewGetRawTransactionCmd(txHash string, verbose *int) *GetRawTransactionCmd 
 	}
 }
 
+// GetStakeDifficultyCmd is a type handling custom marshaling and
+// unmarshaling of getstakedifficulty JSON RPC commands.
+type GetStakeDifficultyCmd struct{}
+
+// NewGetStakeDifficultyCmd returns a new instance which can be used to
+// issue a JSON-RPC getstakedifficulty command.
+func NewGetStakeDifficultyCmd() *GetStakeDifficultyCmd {
+	return &GetStakeDifficultyCmd{}
+}
+
+// GetStakeVersionInfoCmd returns stake version info for the current interval.
+// Optionally, Count indicates how many additional intervals to return.
+type GetStakeVersionInfoCmd struct {
+	Count *int32
+}
+
+// NewGetStakeVersionInfoCmd returns a new instance which can be used to
+// issue a JSON-RPC getstakeversioninfo command.
+func NewGetStakeVersionInfoCmd(count int32) *GetStakeVersionInfoCmd {
+	return &GetStakeVersionInfoCmd{
+		Count: &count,
+	}
+}
+
+// GetStakeVersionsCmd returns stake version for a range of blocks.
+// Count indicates how many blocks are walked backwards.
+type GetStakeVersionsCmd struct {
+	Hash  string
+	Count int32
+}
+
+// NewGetStakeVersionsCmd returns a new instance which can be used to
+// issue a JSON-RPC getstakeversions command.
+func NewGetStakeVersionsCmd(hash string, count int32) *GetStakeVersionsCmd {
+	return &GetStakeVersionsCmd{
+		Hash:  hash,
+		Count: count,
+	}
+}
+
+// GetTicketPoolValueCmd defines the getticketpoolvalue JSON-RPC command.
+type GetTicketPoolValueCmd struct{}
+
+// NewGetTicketPoolValueCmd returns a new instance which can be used to issue a
+// getticketpoolvalue JSON-RPC command.
+func NewGetTicketPoolValueCmd() *GetTicketPoolValueCmd {
+	return &GetTicketPoolValueCmd{}
+}
+
 // GetTxOutCmd defines the gettxout JSON-RPC command.
 type GetTxOutCmd struct {
 	Txid           string
@@ -589,6 +812,20 @@ type GetTxOutSetInfoCmd struct{}
 // gettxoutsetinfo JSON-RPC command.
 func NewGetTxOutSetInfoCmd() *GetTxOutSetInfoCmd {
 	return &GetTxOutSetInfoCmd{}
+}
+
+// GetVoteInfoCmd returns voting results over a range of blocks.  Count
+// indicates how many blocks are walked backwards.
+type GetVoteInfoCmd struct {
+	Version uint32
+}
+
+// NewGetVoteInfoCmd returns a new instance which can be used to
+// issue a JSON-RPC getvoteinfo command.
+func NewGetVoteInfoCmd(version uint32) *GetVoteInfoCmd {
+	return &GetVoteInfoCmd{
+		Version: version,
+	}
 }
 
 // GetWorkCmd defines the getwork JSON-RPC command.
@@ -623,6 +860,46 @@ func NewHelpCmd(command *string) *HelpCmd {
 	}
 }
 
+// LiveTicketsCmd is a type handling custom marshaling and
+// unmarshaling of livetickets JSON RPC commands.
+type LiveTicketsCmd struct{}
+
+// NewLiveTicketsCmd returns a new instance which can be used to issue a JSON-RPC
+// livetickets command.
+func NewLiveTicketsCmd() *LiveTicketsCmd {
+	return &LiveTicketsCmd{}
+}
+
+// MissedTicketsCmd is a type handling custom marshaling and
+// unmarshaling of missedtickets JSON RPC commands.
+type MissedTicketsCmd struct{}
+
+// NewMissedTicketsCmd returns a new instance which can be used to issue a JSON-RPC
+// missedtickets command.
+func NewMissedTicketsCmd() *MissedTicketsCmd {
+	return &MissedTicketsCmd{}
+}
+
+// NodeCmd defines the dropnode JSON-RPC command.
+type NodeCmd struct {
+	SubCmd        NodeSubCmd `jsonrpcusage:"\"connect|remove|disconnect\""`
+	Target        string
+	ConnectSubCmd *string `jsonrpcusage:"\"perm|temp\""`
+}
+
+// NewNodeCmd returns a new instance which can be used to issue a `node`
+// JSON-RPC command.
+//
+// The parameters which are pointers indicate they are optional.  Passing nil
+// for optional parameters will use the default value.
+func NewNodeCmd(subCmd NodeSubCmd, target string, connectSubCmd *string) *NodeCmd {
+	return &NodeCmd{
+		SubCmd:        subCmd,
+		Target:        target,
+		ConnectSubCmd: connectSubCmd,
+	}
+}
+
 // PingCmd defines the ping JSON-RPC command.
 type PingCmd struct{}
 
@@ -630,6 +907,26 @@ type PingCmd struct{}
 // command.
 func NewPingCmd() *PingCmd {
 	return &PingCmd{}
+}
+
+// RebroadcastMissedCmd is a type handling custom marshaling and
+// unmarshaling of rebroadcastwinners JSON RPC commands.
+type RebroadcastMissedCmd struct{}
+
+// NewRebroadcastMissedCmd returns a new instance which can be used to
+// issue a JSON-RPC rebroadcastmissed command.
+func NewRebroadcastMissedCmd() *RebroadcastMissedCmd {
+	return &RebroadcastMissedCmd{}
+}
+
+// RebroadcastWinnersCmd is a type handling custom marshaling and
+// unmarshaling of rebroadcastwinners JSON RPC commands.
+type RebroadcastWinnersCmd struct{}
+
+// NewRebroadcastWinnersCmd returns a new instance which can be used to
+// issue a JSON-RPC rebroadcastwinners command.
+func NewRebroadcastWinnersCmd() *RebroadcastWinnersCmd {
+	return &RebroadcastWinnersCmd{}
 }
 
 // SearchRawTransactionsCmd defines the searchrawtransactions JSON-RPC command.
@@ -730,6 +1027,64 @@ func NewSubmitBlockCmd(hexBlock string, options *SubmitBlockOptions) *SubmitBloc
 	}
 }
 
+// TicketFeeInfoCmd defines the ticketsfeeinfo JSON-RPC command.
+type TicketFeeInfoCmd struct {
+	Blocks  *uint32
+	Windows *uint32
+}
+
+// NewTicketFeeInfoCmd returns a new instance which can be used to issue a
+// JSON-RPC ticket fee info command.
+func NewTicketFeeInfoCmd(blocks *uint32, windows *uint32) *TicketFeeInfoCmd {
+	return &TicketFeeInfoCmd{
+		Blocks:  blocks,
+		Windows: windows,
+	}
+}
+
+// TicketsForAddressCmd defines the ticketsforbucket JSON-RPC command.
+type TicketsForAddressCmd struct {
+	Address string
+}
+
+// NewTicketsForAddressCmd returns a new instance which can be used to issue a
+// JSON-RPC tickets for bucket command.
+func NewTicketsForAddressCmd(addr string) *TicketsForAddressCmd {
+	return &TicketsForAddressCmd{addr}
+}
+
+// TicketVWAPCmd defines the ticketvwap JSON-RPC command.
+type TicketVWAPCmd struct {
+	Start *uint32
+	End   *uint32
+}
+
+// NewTicketVWAPCmd returns a new instance which can be used to issue a
+// JSON-RPC ticket volume weight average price command.
+func NewTicketVWAPCmd(start *uint32, end *uint32) *TicketVWAPCmd {
+	return &TicketVWAPCmd{
+		Start: start,
+		End:   end,
+	}
+}
+
+// TxFeeInfoCmd defines the ticketsfeeinfo JSON-RPC command.
+type TxFeeInfoCmd struct {
+	Blocks     *uint32
+	RangeStart *uint32
+	RangeEnd   *uint32
+}
+
+// NewTxFeeInfoCmd returns a new instance which can be used to issue a
+// JSON-RPC ticket fee info command.
+func NewTxFeeInfoCmd(blocks *uint32, start *uint32, end *uint32) *TxFeeInfoCmd {
+	return &TxFeeInfoCmd{
+		Blocks:     blocks,
+		RangeStart: start,
+		RangeEnd:   end,
+	}
+}
+
 // ValidateAddressCmd defines the validateaddress JSON-RPC command.
 type ValidateAddressCmd struct {
 	Address string
@@ -778,17 +1133,35 @@ func NewVerifyMessageCmd(address, signature, message string) *VerifyMessageCmd {
 	}
 }
 
+// VersionCmd defines the version JSON-RPC command.
+type VersionCmd struct{}
+
+// NewVersionCmd returns a new instance which can be used to issue a JSON-RPC
+// version command.
+func NewVersionCmd() *VersionCmd { return new(VersionCmd) }
+
 func init() {
 	// No special flags for commands in this file.
 	flags := UsageFlag(0)
 
 	MustRegisterCmd("addnode", (*AddNodeCmd)(nil), flags)
 	MustRegisterCmd("createrawtransaction", (*CreateRawTransactionCmd)(nil), flags)
+	MustRegisterCmd("debuglevel", (*DebugLevelCmd)(nil), flags)
 	MustRegisterCmd("decoderawtransaction", (*DecodeRawTransactionCmd)(nil), flags)
 	MustRegisterCmd("decodescript", (*DecodeScriptCmd)(nil), flags)
 	MustRegisterCmd("estimatefee", (*EstimateFeeCmd)(nil), flags)
 	MustRegisterCmd("estimatesmartfee", (*EstimateSmartFeeCmd)(nil), flags)
+	MustRegisterCmd("estimatestakediff", (*EstimateStakeDiffCmd)(nil), flags)
+	MustRegisterCmd("existsaddress", (*ExistsAddressCmd)(nil), flags)
+	MustRegisterCmd("existsaddresses", (*ExistsAddressesCmd)(nil), flags)
+	MustRegisterCmd("existsmissedtickets", (*ExistsMissedTicketsCmd)(nil), flags)
+	MustRegisterCmd("existsexpiredtickets", (*ExistsExpiredTicketsCmd)(nil), flags)
+	MustRegisterCmd("existsliveticket", (*ExistsLiveTicketCmd)(nil), flags)
+	MustRegisterCmd("existslivetickets", (*ExistsLiveTicketsCmd)(nil), flags)
+	MustRegisterCmd("existsmempooltxs", (*ExistsMempoolTxsCmd)(nil), flags)
+	MustRegisterCmd("generate", (*GenerateCmd)(nil), flags)
 	MustRegisterCmd("getaddednodeinfo", (*GetAddedNodeInfoCmd)(nil), flags)
+	MustRegisterCmd("getbestblock", (*GetBestBlockCmd)(nil), flags)
 	MustRegisterCmd("getbestblockhash", (*GetBestBlockHashCmd)(nil), flags)
 	MustRegisterCmd("getblock", (*GetBlockCmd)(nil), flags)
 	MustRegisterCmd("getblockchaininfo", (*GetBlockChainInfoCmd)(nil), flags)
@@ -800,7 +1173,9 @@ func init() {
 	MustRegisterCmd("getcfilter", (*GetCFilterCmd)(nil), flags)
 	MustRegisterCmd("getcfilterheader", (*GetCFilterHeaderCmd)(nil), flags)
 	MustRegisterCmd("getchaintips", (*GetChainTipsCmd)(nil), flags)
+	MustRegisterCmd("getcoinsupply", (*GetCoinSupplyCmd)(nil), flags)
 	MustRegisterCmd("getconnectioncount", (*GetConnectionCountCmd)(nil), flags)
+	MustRegisterCmd("getcurrentnet", (*GetCurrentNetCmd)(nil), flags)
 	MustRegisterCmd("getdifficulty", (*GetDifficultyCmd)(nil), flags)
 	MustRegisterCmd("getgenerate", (*GetGenerateCmd)(nil), flags)
 	MustRegisterCmd("gethashespersec", (*GetHashesPerSecCmd)(nil), flags)
@@ -814,17 +1189,32 @@ func init() {
 	MustRegisterCmd("getpeerinfo", (*GetPeerInfoCmd)(nil), flags)
 	MustRegisterCmd("getrawmempool", (*GetRawMempoolCmd)(nil), flags)
 	MustRegisterCmd("getrawtransaction", (*GetRawTransactionCmd)(nil), flags)
+	MustRegisterCmd("getstakedifficulty", (*GetStakeDifficultyCmd)(nil), flags)
+	MustRegisterCmd("getstakeversioninfo", (*GetStakeVersionInfoCmd)(nil), flags)
+	MustRegisterCmd("getstakeversions", (*GetStakeVersionsCmd)(nil), flags)
+	MustRegisterCmd("getticketpoolvalue", (*GetTicketPoolValueCmd)(nil), flags)
 	MustRegisterCmd("gettxout", (*GetTxOutCmd)(nil), flags)
 	MustRegisterCmd("gettxoutsetinfo", (*GetTxOutSetInfoCmd)(nil), flags)
+	MustRegisterCmd("getvoteinfo", (*GetVoteInfoCmd)(nil), flags)
 	MustRegisterCmd("getwork", (*GetWorkCmd)(nil), flags)
 	MustRegisterCmd("help", (*HelpCmd)(nil), flags)
+	MustRegisterCmd("livetickets", (*LiveTicketsCmd)(nil), flags)
+	MustRegisterCmd("missedtickets", (*MissedTicketsCmd)(nil), flags)
+	MustRegisterCmd("node", (*NodeCmd)(nil), flags)
 	MustRegisterCmd("ping", (*PingCmd)(nil), flags)
+	MustRegisterCmd("rebroadcastmissed", (*RebroadcastMissedCmd)(nil), flags)
+	MustRegisterCmd("rebroadcastwinners", (*RebroadcastWinnersCmd)(nil), flags)
 	MustRegisterCmd("searchrawtransactions", (*SearchRawTransactionsCmd)(nil), flags)
 	MustRegisterCmd("sendrawtransaction", (*SendRawTransactionCmd)(nil), flags)
 	MustRegisterCmd("setgenerate", (*SetGenerateCmd)(nil), flags)
 	MustRegisterCmd("stop", (*StopCmd)(nil), flags)
 	MustRegisterCmd("submitblock", (*SubmitBlockCmd)(nil), flags)
+	MustRegisterCmd("ticketfeeinfo", (*TicketFeeInfoCmd)(nil), flags)
+	MustRegisterCmd("ticketsforaddress", (*TicketsForAddressCmd)(nil), flags)
+	MustRegisterCmd("ticketvwap", (*TicketVWAPCmd)(nil), flags)
+	MustRegisterCmd("txfeeinfo", (*TxFeeInfoCmd)(nil), flags)
 	MustRegisterCmd("validateaddress", (*ValidateAddressCmd)(nil), flags)
 	MustRegisterCmd("verifychain", (*VerifyChainCmd)(nil), flags)
 	MustRegisterCmd("verifymessage", (*VerifyMessageCmd)(nil), flags)
+	MustRegisterCmd("version", (*VersionCmd)(nil), flags)
 }
