+++ dcrjson/chainsvrresults.go
@@ -7,31 +7,54 @@ package dcrjson
 
 import "encoding/json"
 
-// GetBlockHeaderVerboseResult models the data from the getblockheader command when
-// the verbose flag is set.  When the verbose flag is not set, getblockheader
-// returns a hex-encoded string.
-type GetBlockHeaderVerboseResult struct {
-	Hash          string  `json:"hash"`
-	Confirmations int64   `json:"confirmations"`
-	Version       int32   `json:"version"`
-	PreviousHash  string  `json:"previousblockhash,omitempty"`
-	MerkleRoot    string  `json:"merkleroot"`
-	StakeRoot     string  `json:"stakeroot"`
-	VoteBits      uint16  `json:"votebits"`
-	FinalState    string  `json:"finalstate"`
-	Voters        uint16  `json:"voters"`
-	FreshStake    uint8   `json:"freshstake"`
-	Revocations   uint8   `json:"revocations"`
-	PoolSize      uint32  `json:"poolsize"`
-	Bits          string  `json:"bits"`
-	SBits         float64 `json:"sbits"`
-	Height        uint32  `json:"height"`
-	Size          uint32  `json:"size"`
-	Time          int64   `json:"time"`
-	Nonce         uint32  `json:"nonce"`
-	StakeVersion  uint32  `json:"stakeversion"`
-	Difficulty    float64 `json:"difficulty"`
-	NextHash      string  `json:"nextblockhash,omitempty"`
+// TxRawDecodeResult models the data from the decoderawtransaction command.
+type TxRawDecodeResult struct {
+	Txid     string `json:"txid"`
+	Version  int32  `json:"version"`
+	Locktime uint32 `json:"locktime"`
+	Expiry   uint32 `json:"expiry"`
+	Vin      []Vin  `json:"vin"`
+	Vout     []Vout `json:"vout"`
+}
+
+// DecodeScriptResult models the data returned from the decodescript command.
+type DecodeScriptResult struct {
+	Asm       string   `json:"asm"`
+	ReqSigs   int32    `json:"reqSigs,omitempty"`
+	Type      string   `json:"type"`
+	Addresses []string `json:"addresses,omitempty"`
+	P2sh      string   `json:"p2sh,omitempty"`
+}
+
+// EstimateSmartFeeResult models the data returned from the estimatesmartfee
+// command.
+type EstimateSmartFeeResult struct {
+	FeeRate float64  `json:"feerate"`
+	Errors  []string `json:"errors"`
+	Blocks  int64    `json:"blocks"`
+}
+
+// EstimateStakeDiffResult models the data returned from the estimatestakediff
+// command.
+type EstimateStakeDiffResult struct {
+	Min      float64  `json:"min"`
+	Max      float64  `json:"max"`
+	Expected float64  `json:"expected"`
+	User     *float64 `json:"user,omitempty"`
+}
+
+// GetAddedNodeInfoResultAddr models the data of the addresses portion of the
+// getaddednodeinfo command.
+type GetAddedNodeInfoResultAddr struct {
+	Address   string `json:"address"`
+	Connected string `json:"connected"`
+}
+
+// GetAddedNodeInfoResult models the data from the getaddednodeinfo command.
+type GetAddedNodeInfoResult struct {
+	AddedNode string                        `json:"addednode"`
+	Connected *bool                         `json:"connected,omitempty"`
+	Addresses *[]GetAddedNodeInfoResultAddr `json:"addresses,omitempty"`
 }
 
 // GetBlockVerboseResult models the data from the getblock command when the
@@ -66,44 +89,6 @@ type GetBlockVerboseResult struct {
 	NextHash      string        `json:"nextblockhash,omitempty"`
 }
 
-// CreateMultiSigResult models the data returned from the createmultisig
-// command.
-type CreateMultiSigResult struct {
-	Address      string `json:"address"`
-	RedeemScript string `json:"redeemScript"`
-}
-
-// DecodeScriptResult models the data returned from the decodescript command.
-type DecodeScriptResult struct {
-	Asm       string   `json:"asm"`
-	ReqSigs   int32    `json:"reqSigs,omitempty"`
-	Type      string   `json:"type"`
-	Addresses []string `json:"addresses,omitempty"`
-	P2sh      string   `json:"p2sh,omitempty"`
-}
-
-// EstimateSmartFeeResult models the data returned from the estimatesmartfee
-// command.
-type EstimateSmartFeeResult struct {
-	FeeRate float64  `json:"feerate"`
-	Errors  []string `json:"errors"`
-	Blocks  int64    `json:"blocks"`
-}
-
-// GetAddedNodeInfoResultAddr models the data of the addresses portion of the
-// getaddednodeinfo command.
-type GetAddedNodeInfoResultAddr struct {
-	Address   string `json:"address"`
-	Connected string `json:"connected"`
-}
-
-// GetAddedNodeInfoResult models the data from the getaddednodeinfo command.
-type GetAddedNodeInfoResult struct {
-	AddedNode string                        `json:"addednode"`
-	Connected *bool                         `json:"connected,omitempty"`
-	Addresses *[]GetAddedNodeInfoResultAddr `json:"addresses,omitempty"`
-}
-
 // GetBlockChainInfoResult models the data returned from the getblockchaininfo
 // command.
 type GetBlockChainInfoResult struct {
@@ -114,6 +99,33 @@ type GetBlockChainInfoResult struct {
 	Difficulty           float64 `json:"difficulty"`
 	VerificationProgress float64 `json:"verificationprogress"`
 	ChainWork            string  `json:"chainwork"`
+}
+
+// GetBlockHeaderVerboseResult models the data from the getblockheader command when
+// the verbose flag is set.  When the verbose flag is not set, getblockheader
+// returns a hex-encoded string.
+type GetBlockHeaderVerboseResult struct {
+	Hash          string  `json:"hash"`
+	Confirmations int64   `json:"confirmations"`
+	Version       int32   `json:"version"`
+	PreviousHash  string  `json:"previousblockhash,omitempty"`
+	MerkleRoot    string  `json:"merkleroot"`
+	StakeRoot     string  `json:"stakeroot"`
+	VoteBits      uint16  `json:"votebits"`
+	FinalState    string  `json:"finalstate"`
+	Voters        uint16  `json:"voters"`
+	FreshStake    uint8   `json:"freshstake"`
+	Revocations   uint8   `json:"revocations"`
+	PoolSize      uint32  `json:"poolsize"`
+	Bits          string  `json:"bits"`
+	SBits         float64 `json:"sbits"`
+	Height        uint32  `json:"height"`
+	Size          uint32  `json:"size"`
+	Time          int64   `json:"time"`
+	Nonce         uint32  `json:"nonce"`
+	StakeVersion  uint32  `json:"stakeversion"`
+	Difficulty    float64 `json:"difficulty"`
+	NextHash      string  `json:"nextblockhash,omitempty"`
 }
 
 // GetBlockSubsidyResult models the data returned from the getblocksubsidy
@@ -188,11 +200,64 @@ type GetChainTipsResult struct {
 	Status    string `json:"status"`
 }
 
+// GetHeadersResult models the data returned by the chain server getheaders
+// command.
+type GetHeadersResult struct {
+	Headers []string `json:"headers"`
+}
+
+// InfoChainResult models the data returned by the chain server getinfo command.
+type InfoChainResult struct {
+	Version         int32   `json:"version"`
+	ProtocolVersion int32   `json:"protocolversion"`
+	Blocks          int64   `json:"blocks"`
+	TimeOffset      int64   `json:"timeoffset"`
+	Connections     int32   `json:"connections"`
+	Proxy           string  `json:"proxy"`
+	Difficulty      float64 `json:"difficulty"`
+	TestNet         bool    `json:"testnet"`
+	RelayFee        float64 `json:"relayfee"`
+	Errors          string  `json:"errors"`
+}
+
 // GetMempoolInfoResult models the data returned from the getmempoolinfo
 // command.
 type GetMempoolInfoResult struct {
 	Size  int64 `json:"size"`
 	Bytes int64 `json:"bytes"`
+}
+
+// GetMiningInfoResult models the data from the getmininginfo command.
+// Contains Decred additions.
+type GetMiningInfoResult struct {
+	Blocks           int64   `json:"blocks"`
+	CurrentBlockSize uint64  `json:"currentblocksize"`
+	CurrentBlockTx   uint64  `json:"currentblocktx"`
+	Difficulty       float64 `json:"difficulty"`
+	StakeDifficulty  int64   `json:"stakedifficulty"`
+	Errors           string  `json:"errors"`
+	Generate         bool    `json:"generate"`
+	GenProcLimit     int32   `json:"genproclimit"`
+	HashesPerSec     int64   `json:"hashespersec"`
+	NetworkHashPS    int64   `json:"networkhashps"`
+	PooledTx         uint64  `json:"pooledtx"`
+	TestNet          bool    `json:"testnet"`
+}
+
+// LocalAddressesResult models the localaddresses data from the getnetworkinfo
+// command.
+type LocalAddressesResult struct {
+	Address string `json:"address"`
+	Port    uint16 `json:"port"`
+	Score   int32  `json:"score"`
+}
+
+// NetworksResult models the networks data from the getnetworkinfo command.
+type NetworksResult struct {
+	Name      string `json:"name"`
+	Limited   bool   `json:"limited"`
+	Reachable bool   `json:"reachable"`
+	Proxy     string `json:"proxy"`
 }
 
 // GetNetworkInfoResult models the data returned from the getnetworkinfo
@@ -205,6 +270,13 @@ type GetNetworkInfoResult struct {
 	Networks        []NetworksResult       `json:"networks"`
 	RelayFee        float64                `json:"relayfee"`
 	LocalAddresses  []LocalAddressesResult `json:"localaddresses"`
+}
+
+// GetNetTotalsResult models the data returned from the getnettotals command.
+type GetNetTotalsResult struct {
+	TotalBytesRecv uint64 `json:"totalbytesrecv"`
+	TotalBytesSent uint64 `json:"totalbytessent"`
+	TimeMillis     int64  `json:"timemillis"`
 }
 
 // GetPeerInfoResult models the data returned from the getpeerinfo command.
@@ -244,15 +316,71 @@ type GetRawMempoolVerboseResult struct {
 	Depends          []string `json:"depends"`
 }
 
-// ScriptPubKeyResult models the scriptPubKey data of a tx script.  It is
-// defined separately since it is used by multiple commands.
-type ScriptPubKeyResult struct {
-	Asm       string   `json:"asm"`
-	Hex       string   `json:"hex,omitempty"`
-	ReqSigs   int32    `json:"reqSigs,omitempty"`
-	Type      string   `json:"type"`
-	Addresses []string `json:"addresses,omitempty"`
-	CommitAmt *float64 `json:"commitamt,omitempty"`
+// TxRawResult models the data from the getrawtransaction command.
+type TxRawResult struct {
+	Hex           string `json:"hex"`
+	Txid          string `json:"txid"`
+	Version       int32  `json:"version"`
+	LockTime      uint32 `json:"locktime"`
+	Expiry        uint32 `json:"expiry"`
+	Vin           []Vin  `json:"vin"`
+	Vout          []Vout `json:"vout"`
+	BlockHash     string `json:"blockhash,omitempty"`
+	BlockHeight   int64  `json:"blockheight"`
+	BlockIndex    uint32 `json:"blockindex,omitempty"`
+	Confirmations int64  `json:"confirmations,omitempty"`
+	Time          int64  `json:"time,omitempty"`
+	Blocktime     int64  `json:"blocktime,omitempty"`
+}
+
+// GetStakeDifficultyResult models the data returned from the
+// getstakedifficulty command.
+type GetStakeDifficultyResult struct {
+	CurrentStakeDifficulty float64 `json:"current"`
+	NextStakeDifficulty    float64 `json:"next"`
+}
+
+// VersionCount models a generic version:count tuple.
+type VersionCount struct {
+	Version uint32 `json:"version"`
+	Count   uint32 `json:"count"`
+}
+
+// VersionInterval models a cooked version count for an interval.
+type VersionInterval struct {
+	StartHeight  int64          `json:"startheight"`
+	EndHeight    int64          `json:"endheight"`
+	PoSVersions  []VersionCount `json:"posversions"`
+	VoteVersions []VersionCount `json:"voteversions"`
+}
+
+// GetStakeVersionInfoResult models the resulting data for getstakeversioninfo
+// command.
+type GetStakeVersionInfoResult struct {
+	CurrentHeight int64             `json:"currentheight"`
+	Hash          string            `json:"hash"`
+	Intervals     []VersionInterval `json:"intervals"`
+}
+
+// VersionBits models a generic version:bits tuple.
+type VersionBits struct {
+	Version uint32 `json:"version"`
+	Bits    uint16 `json:"bits"`
+}
+
+// StakeVersions models the data for GetStakeVersionsResult.
+type StakeVersions struct {
+	Hash         string        `json:"hash"`
+	Height       int64         `json:"height"`
+	BlockVersion int32         `json:"blockversion"`
+	StakeVersion uint32        `json:"stakeversion"`
+	Votes        []VersionBits `json:"votes"`
+}
+
+// GetStakeVersionsResult models the data returned from the getstakeversions
+// command.
+type GetStakeVersionsResult struct {
+	StakeVersions []StakeVersions `json:"stakeversions"`
 }
 
 // GetTxOutResult models the data from the gettxout command.
@@ -265,11 +393,172 @@ type GetTxOutResult struct {
 	Coinbase      bool               `json:"coinbase"`
 }
 
-// GetNetTotalsResult models the data returned from the getnettotals command.
-type GetNetTotalsResult struct {
-	TotalBytesRecv uint64 `json:"totalbytesrecv"`
-	TotalBytesSent uint64 `json:"totalbytessent"`
-	TimeMillis     int64  `json:"timemillis"`
+// Choice models an individual choice inside an Agenda.
+type Choice struct {
+	Id          string  `json:"id"`
+	Description string  `json:"description"`
+	Bits        uint16  `json:"bits"`
+	IsAbstain   bool    `json:"isabstain"`
+	IsNo        bool    `json:"isno"`
+	Count       uint32  `json:"count"`
+	Progress    float64 `json:"progress"`
+}
+
+// Agenda models an individual agenda including its choices.
+type Agenda struct {
+	Id             string   `json:"id"`
+	Description    string   `json:"description"`
+	Mask           uint16   `json:"mask"`
+	StartTime      uint64   `json:"starttime"`
+	ExpireTime     uint64   `json:"expiretime"`
+	Status         string   `json:"status"`
+	QuorumProgress float64  `json:"quorumprogress"`
+	Choices        []Choice `json:"choices"`
+}
+
+// GetVoteInfoResult models the data returned from the getvoteinfo command.
+type GetVoteInfoResult struct {
+	CurrentHeight int64    `json:"currentheight"`
+	StartHeight   int64    `json:"startheight"`
+	EndHeight     int64    `json:"endheight"`
+	Hash          string   `json:"hash"`
+	VoteVersion   uint32   `json:"voteversion"`
+	Quorum        uint32   `json:"quorum"`
+	TotalVotes    uint32   `json:"totalvotes"`
+	Agendas       []Agenda `json:"agendas,omitempty"`
+}
+
+// GetWorkResult models the data from the getwork command.
+type GetWorkResult struct {
+	Data   string `json:"data"`
+	Target string `json:"target"`
+}
+
+// Ticket is the structure representing a ticket.
+type Ticket struct {
+	Hash  string `json:"hash"`
+	Owner string `json:"owner"`
+}
+
+// LiveTicketsResult models the data returned from the livetickets
+// command.
+type LiveTicketsResult struct {
+	Tickets []string `json:"tickets"`
+}
+
+// MissedTicketsResult models the data returned from the missedtickets
+// command.
+type MissedTicketsResult struct {
+	Tickets []string `json:"tickets"`
+}
+
+// FeeInfoBlock is ticket fee information about a block.
+type FeeInfoBlock struct {
+	Height uint32  `json:"height"`
+	Number uint32  `json:"number"`
+	Min    float64 `json:"min"`
+	Max    float64 `json:"max"`
+	Mean   float64 `json:"mean"`
+	Median float64 `json:"median"`
+	StdDev float64 `json:"stddev"`
+}
+
+// FeeInfoMempool is ticket fee information about the mempool.
+type FeeInfoMempool struct {
+	Number uint32  `json:"number"`
+	Min    float64 `json:"min"`
+	Max    float64 `json:"max"`
+	Mean   float64 `json:"mean"`
+	Median float64 `json:"median"`
+	StdDev float64 `json:"stddev"`
+}
+
+// FeeInfoRange is ticket fee information about a range.
+type FeeInfoRange struct {
+	Number uint32  `json:"number"`
+	Min    float64 `json:"min"`
+	Max    float64 `json:"max"`
+	Mean   float64 `json:"mean"`
+	Median float64 `json:"median"`
+	StdDev float64 `json:"stddev"`
+}
+
+// FeeInfoWindow is ticket fee information about an adjustment window.
+type FeeInfoWindow struct {
+	StartHeight uint32  `json:"startheight"`
+	EndHeight   uint32  `json:"endheight"`
+	Number      uint32  `json:"number"`
+	Min         float64 `json:"min"`
+	Max         float64 `json:"max"`
+	Mean        float64 `json:"mean"`
+	Median      float64 `json:"median"`
+	StdDev      float64 `json:"stddev"`
+}
+
+// TicketFeeInfoResult models the data returned from the ticketfeeinfo command.
+// command.
+type TicketFeeInfoResult struct {
+	FeeInfoMempool FeeInfoMempool  `json:"feeinfomempool"`
+	FeeInfoBlocks  []FeeInfoBlock  `json:"feeinfoblocks"`
+	FeeInfoWindows []FeeInfoWindow `json:"feeinfowindows"`
+}
+
+// SearchRawTransactionsResult models the data from the searchrawtransaction
+// command.
+type SearchRawTransactionsResult struct {
+	Hex           string       `json:"hex,omitempty"`
+	Txid          string       `json:"txid"`
+	Version       int32        `json:"version"`
+	LockTime      uint32       `json:"locktime"`
+	Vin           []VinPrevOut `json:"vin"`
+	Vout          []Vout       `json:"vout"`
+	BlockHash     string       `json:"blockhash,omitempty"`
+	Confirmations uint64       `json:"confirmations,omitempty"`
+	Time          int64        `json:"time,omitempty"`
+	Blocktime     int64        `json:"blocktime,omitempty"`
+}
+
+// TxFeeInfoResult models the data returned from the ticketfeeinfo command.
+// command.
+type TxFeeInfoResult struct {
+	FeeInfoMempool FeeInfoMempool `json:"feeinfomempool"`
+	FeeInfoBlocks  []FeeInfoBlock `json:"feeinfoblocks"`
+	FeeInfoRange   FeeInfoRange   `json:"feeinforange"`
+}
+
+// TicketsForAddressResult models the data returned from the ticketforaddress
+// command.
+type TicketsForAddressResult struct {
+	Tickets []string `json:"tickets"`
+}
+
+// ValidateAddressChainResult models the data returned by the chain server
+// validateaddress command.
+type ValidateAddressChainResult struct {
+	IsValid bool   `json:"isvalid"`
+	Address string `json:"address,omitempty"`
+}
+
+// VersionResult models objects included in the version response.  In the actual
+// result, these objects are keyed by the program or API name.
+type VersionResult struct {
+	VersionString string `json:"versionstring"`
+	Major         uint32 `json:"major"`
+	Minor         uint32 `json:"minor"`
+	Patch         uint32 `json:"patch"`
+	Prerelease    string `json:"prerelease"`
+	BuildMetadata string `json:"buildmetadata"`
+}
+
+// ScriptPubKeyResult models the scriptPubKey data of a tx script.  It is
+// defined separately since it is used by multiple commands.
+type ScriptPubKeyResult struct {
+	Asm       string   `json:"asm"`
+	Hex       string   `json:"hex,omitempty"`
+	ReqSigs   int32    `json:"reqSigs,omitempty"`
+	Type      string   `json:"type"`
+	Addresses []string `json:"addresses,omitempty"`
+	CommitAmt *float64 `json:"commitamt,omitempty"`
 }
 
 // ScriptSig models a signature script.  It is defined separately since it only
@@ -454,112 +743,4 @@ type Vout struct {
 	N            uint32             `json:"n"`
 	Version      uint16             `json:"version"`
 	ScriptPubKey ScriptPubKeyResult `json:"scriptPubKey"`
-}
-
-// GetMiningInfoResult models the data from the getmininginfo command.
-// Contains Decred additions.
-type GetMiningInfoResult struct {
-	Blocks           int64   `json:"blocks"`
-	CurrentBlockSize uint64  `json:"currentblocksize"`
-	CurrentBlockTx   uint64  `json:"currentblocktx"`
-	Difficulty       float64 `json:"difficulty"`
-	StakeDifficulty  int64   `json:"stakedifficulty"`
-	Errors           string  `json:"errors"`
-	Generate         bool    `json:"generate"`
-	GenProcLimit     int32   `json:"genproclimit"`
-	HashesPerSec     int64   `json:"hashespersec"`
-	NetworkHashPS    int64   `json:"networkhashps"`
-	PooledTx         uint64  `json:"pooledtx"`
-	TestNet          bool    `json:"testnet"`
-}
-
-// GetWorkResult models the data from the getwork command.
-type GetWorkResult struct {
-	Data   string `json:"data"`
-	Target string `json:"target"`
-}
-
-// InfoChainResult models the data returned by the chain server getinfo command.
-type InfoChainResult struct {
-	Version         int32   `json:"version"`
-	ProtocolVersion int32   `json:"protocolversion"`
-	Blocks          int64   `json:"blocks"`
-	TimeOffset      int64   `json:"timeoffset"`
-	Connections     int32   `json:"connections"`
-	Proxy           string  `json:"proxy"`
-	Difficulty      float64 `json:"difficulty"`
-	TestNet         bool    `json:"testnet"`
-	RelayFee        float64 `json:"relayfee"`
-	Errors          string  `json:"errors"`
-}
-
-// LocalAddressesResult models the localaddresses data from the getnetworkinfo
-// command.
-type LocalAddressesResult struct {
-	Address string `json:"address"`
-	Port    uint16 `json:"port"`
-	Score   int32  `json:"score"`
-}
-
-// NetworksResult models the networks data from the getnetworkinfo command.
-type NetworksResult struct {
-	Name      string `json:"name"`
-	Limited   bool   `json:"limited"`
-	Reachable bool   `json:"reachable"`
-	Proxy     string `json:"proxy"`
-}
-
-// TxRawResult models the data from the getrawtransaction command.
-type TxRawResult struct {
-	Hex           string `json:"hex"`
-	Txid          string `json:"txid"`
-	Version       int32  `json:"version"`
-	LockTime      uint32 `json:"locktime"`
-	Expiry        uint32 `json:"expiry"`
-	Vin           []Vin  `json:"vin"`
-	Vout          []Vout `json:"vout"`
-	BlockHash     string `json:"blockhash,omitempty"`
-	BlockHeight   int64  `json:"blockheight"`
-	BlockIndex    uint32 `json:"blockindex,omitempty"`
-	Confirmations int64  `json:"confirmations,omitempty"`
-	Time          int64  `json:"time,omitempty"`
-	Blocktime     int64  `json:"blocktime,omitempty"`
-}
-
-// SearchRawTransactionsResult models the data from the searchrawtransaction
-// command.
-type SearchRawTransactionsResult struct {
-	Hex           string       `json:"hex,omitempty"`
-	Txid          string       `json:"txid"`
-	Version       int32        `json:"version"`
-	LockTime      uint32       `json:"locktime"`
-	Vin           []VinPrevOut `json:"vin"`
-	Vout          []Vout       `json:"vout"`
-	BlockHash     string       `json:"blockhash,omitempty"`
-	Confirmations uint64       `json:"confirmations,omitempty"`
-	Time          int64        `json:"time,omitempty"`
-	Blocktime     int64        `json:"blocktime,omitempty"`
-}
-
-// TxRawDecodeResult models the data from the decoderawtransaction command.
-type TxRawDecodeResult struct {
-	Txid     string `json:"txid"`
-	Version  int32  `json:"version"`
-	Locktime uint32 `json:"locktime"`
-	Expiry   uint32 `json:"expiry"`
-	Vin      []Vin  `json:"vin"`
-	Vout     []Vout `json:"vout"`
-}
-
-// ValidateAddressChainResult models the data returned by the chain server
-// validateaddress command.
-type ValidateAddressChainResult struct {
-	IsValid bool   `json:"isvalid"`
-	Address string `json:"address,omitempty"`
-}
-
-// GetHeadersResult models the data returned by the chain server getheaders
-// command.
-type GetHeadersResult struct {
-	Headers []string `json:"headers"`
 }
