--- dcrjson/dcrdcmds_test.go
@@ -1,139 +1,0 @@
-// Copyright (c) 2015-2016 The Decred developers
-// Use of this source code is governed by an ISC
-// license that can be found in the LICENSE file.
-
-package dcrjson_test
-
-import (
-	"bytes"
-	"encoding/json"
-	"fmt"
-	"reflect"
-	"testing"
-
-	"github.com/decred/dcrd/dcrjson"
-)
-
-// TestBtcdCmds tests all of the btcd extended commands marshal and unmarshal
-// into valid results include handling of optional fields being omitted in the
-// marshalled command, while optional fields with defaults have the default
-// assigned on unmarshalled commands.
-func TestDcrdCmds(t *testing.T) {
-	t.Parallel()
-
-	testID := int(1)
-	tests := []struct {
-		name         string
-		newCmd       func() (interface{}, error)
-		staticCmd    func() interface{}
-		marshalled   string
-		unmarshalled interface{}
-	}{
-		{
-			name: "debuglevel",
-			newCmd: func() (interface{}, error) {
-				return dcrjson.NewCmd("debuglevel", "trace")
-			},
-			staticCmd: func() interface{} {
-				return dcrjson.NewDebugLevelCmd("trace")
-			},
-			marshalled: `{"jsonrpc":"1.0","method":"debuglevel","params":["trace"],"id":1}`,
-			unmarshalled: &dcrjson.DebugLevelCmd{
-				LevelSpec: "trace",
-			},
-		},
-		{
-			name: "getstakeversions",
-			newCmd: func() (interface{}, error) {
-				return dcrjson.NewCmd("getstakeversions", "deadbeef", 1)
-			},
-			staticCmd: func() interface{} {
-				return dcrjson.NewGetStakeVersionsCmd("deadbeef", 1)
-			},
-			marshalled: `{"jsonrpc":"1.0","method":"getstakeversions","params":["deadbeef",1],"id":1}`,
-			unmarshalled: &dcrjson.GetStakeVersionsCmd{
-				Hash:  "deadbeef",
-				Count: 1,
-			},
-		},
-		{
-			name: "getvoteinfo",
-			newCmd: func() (interface{}, error) {
-				return dcrjson.NewCmd("getvoteinfo", 1)
-			},
-			staticCmd: func() interface{} {
-				return dcrjson.NewGetVoteInfoCmd(1)
-			},
-			marshalled: `{"jsonrpc":"1.0","method":"getvoteinfo","params":[1],"id":1}`,
-			unmarshalled: &dcrjson.GetVoteInfoCmd{
-				Version: 1,
-			},
-		},
-	}
-
-	t.Logf("Running %d tests", len(tests))
-	for i, test := range tests {
-		// Marshal the command as created by the new static command
-		// creation function.
-		marshalled, err := dcrjson.MarshalCmd("1.0", testID, test.staticCmd())
-		if err != nil {
-			t.Errorf("MarshalCmd #%d (%s) unexpected error: %v", i,
-				test.name, err)
-			continue
-		}
-
-		if !bytes.Equal(marshalled, []byte(test.marshalled)) {
-			t.Errorf("Test #%d (%s) unexpected marshalled data - "+
-				"got %s, want %s", i, test.name, marshalled,
-				test.marshalled)
-			continue
-		}
-
-		// Ensure the command is created without error via the generic
-		// new command creation function.
-		cmd, err := test.newCmd()
-		if err != nil {
-			t.Errorf("Test #%d (%s) unexpected NewCmd error: %v ",
-				i, test.name, err)
-		}
-
-		// Marshal the command as created by the generic new command
-		// creation function.
-		marshalled, err = dcrjson.MarshalCmd("1.0", testID, cmd)
-		if err != nil {
-			t.Errorf("MarshalCmd #%d (%s) unexpected error: %v", i,
-				test.name, err)
-			continue
-		}
-
-		if !bytes.Equal(marshalled, []byte(test.marshalled)) {
-			t.Errorf("Test #%d (%s) unexpected marshalled data - "+
-				"got %s, want %s", i, test.name, marshalled,
-				test.marshalled)
-			continue
-		}
-
-		var request dcrjson.Request
-		if err := json.Unmarshal(marshalled, &request); err != nil {
-			t.Errorf("Test #%d (%s) unexpected error while "+
-				"unmarshalling JSON-RPC request: %v", i,
-				test.name, err)
-			continue
-		}
-
-		cmd, err = dcrjson.UnmarshalCmd(&request)
-		if err != nil {
-			t.Errorf("UnmarshalCmd #%d (%s) unexpected error: %v", i,
-				test.name, err)
-			continue
-		}
-
-		if !reflect.DeepEqual(cmd, test.unmarshalled) {
-			t.Errorf("Test #%d (%s) unexpected unmarshalled command "+
-				"- got %s, want %s", i, test.name,
-				fmt.Sprintf("(%T) %+[1]v", cmd),
-				fmt.Sprintf("(%T) %+[1]v\n", test.unmarshalled))
-			continue
-		}
-	}
-}
