+++ dcrjson/walletsvrwsntfns_test.go
@@ -50,24 +50,27 @@ func TestWalletSvrWsNtfns(t *testing.T) {
 				return dcrjson.NewCmd("dcrdconnected", true)
 			},
 			staticNtfn: func() interface{} {
-				return dcrjson.NewBtcdConnectedNtfn(true)
+				return dcrjson.NewDcrdConnectedNtfn(true)
 			},
 			marshalled: `{"jsonrpc":"1.0","method":"dcrdconnected","params":[true],"id":null}`,
-			unmarshalled: &dcrjson.BtcdConnectedNtfn{
+			unmarshalled: &dcrjson.DcrdConnectedNtfn{
 				Connected: true,
 			},
 		},
 		{
-			name: "walletlockstate",
+			name: "newtickets",
 			newNtfn: func() (interface{}, error) {
-				return dcrjson.NewCmd("walletlockstate", true)
+				return dcrjson.NewCmd("newtickets", "123", 100, 3, []string{"a", "b"})
 			},
 			staticNtfn: func() interface{} {
-				return dcrjson.NewWalletLockStateNtfn(true)
+				return dcrjson.NewNewTicketsNtfn("123", 100, 3, []string{"a", "b"})
 			},
-			marshalled: `{"jsonrpc":"1.0","method":"walletlockstate","params":[true],"id":null}`,
-			unmarshalled: &dcrjson.WalletLockStateNtfn{
-				Locked: true,
+			marshalled: `{"jsonrpc":"1.0","method":"newtickets","params":["123",100,3,["a","b"]],"id":null}`,
+			unmarshalled: &dcrjson.NewTicketsNtfn{
+				Hash:      "123",
+				Height:    100,
+				StakeDiff: 3,
+				Tickets:   []string{"a", "b"},
 			},
 		},
 		{
@@ -109,6 +112,95 @@ func TestWalletSvrWsNtfns(t *testing.T) {
 					Vout:            789,
 					OtherAccount:    "otheracct",
 				},
+			},
+		},
+		{
+			name: "revocationcreated",
+			newNtfn: func() (interface{}, error) {
+				return dcrjson.NewCmd("revocationcreated", "123", "1234")
+			},
+			staticNtfn: func() interface{} {
+				return dcrjson.NewRevocationCreatedNtfn("123", "1234")
+			},
+			marshalled: `{"jsonrpc":"1.0","method":"revocationcreated","params":["123","1234"],"id":null}`,
+			unmarshalled: &dcrjson.RevocationCreatedNtfn{
+				TxHash: "123",
+				SStxIn: "1234",
+			},
+		},
+		{
+			name: "spentandmissedtickets",
+			newNtfn: func() (interface{}, error) {
+				return dcrjson.NewCmd("spentandmissedtickets", "123", 100, 3, map[string]string{"a": "b"})
+			},
+			staticNtfn: func() interface{} {
+				return dcrjson.NewSpentAndMissedTicketsNtfn("123", 100, 3, map[string]string{"a": "b"})
+			},
+			marshalled: `{"jsonrpc":"1.0","method":"spentandmissedtickets","params":["123",100,3,{"a":"b"}],"id":null}`,
+			unmarshalled: &dcrjson.SpentAndMissedTicketsNtfn{
+				Hash:      "123",
+				Height:    100,
+				StakeDiff: 3,
+				Tickets:   map[string]string{"a": "b"},
+			},
+		},
+		{
+			name: "ticketpurchase",
+			newNtfn: func() (interface{}, error) {
+				return dcrjson.NewCmd("ticketpurchased", "123", 5)
+			},
+			staticNtfn: func() interface{} {
+				return dcrjson.NewTicketPurchasedNtfn("123", 5)
+			},
+			marshalled: `{"jsonrpc":"1.0","method":"ticketpurchased","params":["123",5],"id":null}`,
+			unmarshalled: &dcrjson.TicketPurchasedNtfn{
+				TxHash: "123",
+				Amount: 5,
+			},
+		},
+		{
+			name: "votecreated",
+			newNtfn: func() (interface{}, error) {
+				return dcrjson.NewCmd("votecreated", "123", "1234", 100, "12345", 1)
+			},
+			staticNtfn: func() interface{} {
+				return dcrjson.NewVoteCreatedNtfn("123", "1234", 100, "12345", 1)
+			},
+			marshalled: `{"jsonrpc":"1.0","method":"votecreated","params":["123","1234",100,"12345",1],"id":null}`,
+			unmarshalled: &dcrjson.VoteCreatedNtfn{
+				TxHash:    "123",
+				BlockHash: "1234",
+				Height:    100,
+				SStxIn:    "12345",
+				VoteBits:  1,
+			},
+		},
+		{
+			name: "walletlockstate",
+			newNtfn: func() (interface{}, error) {
+				return dcrjson.NewCmd("walletlockstate", true)
+			},
+			staticNtfn: func() interface{} {
+				return dcrjson.NewWalletLockStateNtfn(true)
+			},
+			marshalled: `{"jsonrpc":"1.0","method":"walletlockstate","params":[true],"id":null}`,
+			unmarshalled: &dcrjson.WalletLockStateNtfn{
+				Locked: true,
+			},
+		},
+		{
+			name: "winningtickets",
+			newNtfn: func() (interface{}, error) {
+				return dcrjson.NewCmd("winningtickets", "123", 100, map[string]string{"a": "b"})
+			},
+			staticNtfn: func() interface{} {
+				return dcrjson.NewWinningTicketsNtfn("123", 100, map[string]string{"a": "b"})
+			},
+			marshalled: `{"jsonrpc":"1.0","method":"winningtickets","params":["123",100,{"a":"b"}],"id":null}`,
+			unmarshalled: &dcrjson.WinningTicketsNtfn{
+				BlockHash:   "123",
+				BlockHeight: 100,
+				Tickets:     map[string]string{"a": "b"},
 			},
 		},
 	}
